Implements custom error type for PackageName parsing
Parsing is a situation where it is often very useful to have custom
error types. The reason for this is that it makes it simpler to write
useful unit tests that catch specific parsing errors. In this commit,
the `PackageNameError` type was introduced which catches the three error
kinds that can occur when parsing a package name.

Additionally, unit tests are added that trigger each of the
aforementioned error cases.

Finally, the `Debug` trait is implemented by derive macro rather than
manually, as there does not seem to be a good reason for doing it
manually.