Changes the `Deref` trait of `PackageName` to target str.
The `Deref` trait allows for types to automatically downcast to another
type. It is one of the few places where the compiler tries to be smart,
rather than requiring the programmer to be explicit.

Usually, this trait is used to make custom types behave like more
generic types. In this case, the `PackageName` type wraps a string, and
should behave mostly like one.

However, the appropriate type to target here is not `String`, but `str`.
Why is that? Both are stringly typed, but there is an important
difference. `str` is the most generic type — it is the equivalent of a
UTF-8 encoded unsized byte array. `String` is one possible container
type for `str`, but not the only one. In fact, the choice of the
container type is arbitrary and an implementation detail.

In the interest of building abstractions but being flexible, any types
which implement `Deref` to some string type should dereference to an
`&str` rather than a `&String`.