Moves resolving code into resolver module
One module should do one thing. The `package` module defines types and
methods to work with packages and the package store. Previously, it also
contained code to resolve package dependencies, but this commit moves
that logic out into it's own module since it is sufficiently separate
and self-contained that in the interest of encapsulation, it should be
considered a separate thing.