Don't use unwrap_or_default in PackageType Display.
There are different kinds of errors that can crop up. Some errors need
to be neatly returned in a way that they can be handled, some errors
need to be passed through to be displayed.

In the case of the `to_str()` method converting a `PackageType` to a
string, this is not something that should reasonable ever fail, because
`PackageType` is an enum. However, due to this API being built such that
it accepts any type that implements `Serialize`, the function returns a
`Result`. The important thing here is that:

- This should never error.
- If it does error, it constitutes a major issue with the code.
- This error cannot be handled, but it must at least be communicated.

Using `unwrap_or_default()` here is a bad idea because it hides the fact
that an error has happened, if it were ever to occur. This is a very
good example of a situation where code should panic. Using the
`unreachable` macro here communicates that this is not something that
will ever error, and the panic is not left in place because someone was
too lazy to handle it properly.