Makes `PackageStore` more robust by not relying on implicit assumptions.
The current implementation of `PackageStore` uses paths that are
resolved using the current working directory. This approach is not a
recommended one, for multiple reasons:

- Application logic should not rely on the current working directory
  staying the same during a program run, as libraries could at any time
  change it. It is implicit global state.
- Relying on this makes it more difficult to thoroughly test the code.
- Relying on implicit global state makes it more difficult to debug some
  code, as the runtime behaviour depends on invisible global variables.

The approach here is that the `PackageStore` now instead contains a
`PathBuf` which represents the root directory from which all relative
paths are resolved. This is robust against changing the current working
directory.

This strategy is also very suitable for unit testing, as it allows
multiple `PackageStore` instances to be used simultaneously with little
effort.