Don't read Manifest within PackageStore::release()
The separation of concerns is a very useful technique that allows
building software that is not too tightly coupled. Software built like
this tends to be more testable, and as a result more robust.

In this case, the `PackageStore` is responsible for one thing, and one
thing only: storing packages. It should not read manifests, compress
data or do anything else that goes beyond what it was intended.

For this reason, a better design is for the `release()` method to accept
a `Manifest` that is already read. That way, this module does not need
to concern itself with how manifests are to be read.