Adapt commands to operate under vendor directory
Since all managed files are automatically copied to the vendor directory
which can be used as a root directory resolving all imports of absolute
paths, this is reflected in the output of ``buffrs ls`` as well.

This allows the compilation of proto files with ``protoc`` like this:

```sh
protoc --proto_path proto/vendor $(buffrs ls) --python_out out/
```

``buffrs lint`` is also adapted to use the vendor directory as root.
You therefore get now errors if you do not use absolute imports.

Given a package ``my-package`` with this structure:

```
proto
├── children
│   └── child.proto
├── root.proto
├── sibling.proto
└── vendor
    ├── my-dependency
    │   └── dep.proto
    └── my-lib
        ├── children
        │   └── child.proto
        ├── root.proto
        └── sibling.proto
```

You want to import ``root.proto`` in ``sibiling.proto``, you have to use
an absolute path with reference to the package name:

```
import "my-lib/root.proto"
```